% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{locate_xtrem}
\alias{locate_xtrem}
\title{Locate xtremas in a series}
\usage{
locate_xtrem(x, last = FALSE)
}
\arguments{
\item{x}{a series of numeric data}

\item{last}{logical value indicating which value (and index) to keep in case of duplicate xtremas,}
}
\value{
A tibble of Min and Max indexes and their corresponding values
}
\description{
Locate xtremas in a series
}
