% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{remove_noise}
\alias{remove_noise}
\title{Remove noise after xtremas detection}
\usage{
remove_noise(xtrem_data, last = FALSE, thresh = 0.05, ...)
}
\arguments{
\item{xtrem_data}{a tibble containing xtrem data}

\item{last}{a logical indicating whether the function should keep the first or last index in case of duplicate xtrem}

\item{thresh}{numeric value in [0;1] representing threshold to detect artefacts}

\item{...}{additional parameters used only in recursive calls}
}
\value{
a clean xtrem data tibble without artefacts
}
\description{
Remove noise after xtremas detection
}
