% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{remove_xtrem_duplicates}
\alias{remove_xtrem_duplicates}
\title{Remove duplicates xtremas}
\usage{
remove_xtrem_duplicates(data, last)
}
\arguments{
\item{data}{xtrem data tibble with potential duplicates}

\item{last}{lagical idincating which value (and index)
to keep in case of duplicate xtremas}
}
\value{
xtrem data tibble without duplicate xtremas
}
\description{
Remove duplicates xtremas
}
